Brave utils: initial commit (#12412)
* Create index.js

* Create nativeAssets.js

* Update braveBidAdapter.js

* Update videoheroesBidAdapter.js

* Update videoheroesBidAdapter.js

* Update braveBidAdapter.js

* Update adotBidAdapter.js

* Update adotBidAdapter.js

* Update adotBidAdapter.js

* Update adotBidAdapter.js

* Update adotBidAdapter.js

* Update adotBidAdapter.js

* Update braveBidAdapter.js

* Update videoheroesBidAdapter.js

* Update videoheroesBidAdapter.js

* Update braveBidAdapter.js

* Update adotBidAdapter.js

* Create buildAndInterpret.js

* Update braveBidAdapter.js

* Update videoheroesBidAdapter.js

* Update buildAndInterpret.js

* Update braveBidAdapter.js

* Update videoheroesBidAdapter.js

* Update braveBidAdapter.js

* Update videoheroesBidAdapter.js

* Update buildAndInterpret.js
